Comments

+++ app/Http/Controllers/ProgramFeeController.php
@@ -110,6 +110,11 @@ class ProgramFeeController extends Controller
         }
 
 
+        if(ProgramFee::isUsed($request->get('campus_program_id'),$request->get('year'))){
+             return redirect()->back()->with('error','Programme fee cannot be updated because it has already been used');
+        }
+
+
         (new ProgramFeeAction)->update($request);
 
         return Util::requestResponse($request,'Program fee updated successfully');
@@ -121,6 +126,9 @@ class ProgramFeeController extends Controller
     public function destroy(Request $request, $id)
     {
         try{
+            if(ProgramFee::isUsed($request->get('campus_program_id'),$request->get('year'))){
+             return redirect()->back()->with('error','Programme fee cannot be deleted because it has already been used');
+            }
             $fee = ProgramFee::findOrFail($id);
             $fee->delete();
             return redirect()->back()->with('message','Program fee deleted successfully');
